Delete LICENSE
I'm deleting the license because now the source is closed; given that there has been no contributions to the existing source code; I'm happy to make this change. I will revisit the licensing when I'm closer to a somewhat stable reason.

This design has been made knowing that I can still maintain the Apache 2.0 licensing for the current state of the project; which is not really to unfortunate given the state of the current application. If there are any legal issues later for making this change I can show that I was the sole contributor when making this change by going through commits.

Deleting this file doesn't mean the file is actually deleted; it's just being replaced as it's available through commit history. Once again, I'm more than happy to accept the Apache 2.0 licensing agreement up until the last commit.